feat: extract HM/Item checking logic from suggestionEngine into encounterTools
Extracted the complex inline HM/Item dependency checking logic and post-processing filtering loop from `suggestionEngine.ts` into a dedicated `encounterTools.ts` utility file. This simplifies the monolithic suggestion generator function while maintaining optimal in-place data mutations.
