+++ .jules/sculptor.md
@@ -18,3 +18,6 @@ Extracted complex inline IIFE math calculations and tailwind styling logic for c
 - **Monolithic Engine Refactoring**: Successfully refactored `src/engine/assistant/suggestionEngine.ts` by splitting out giant logic functions (`generateCatchSuggestions`, `generateGiftAndTradeSuggestions`, `generateBreedingSuggestions`, `generateEvolutionSuggestions`) into dedicated module files under `src/engine/assistant/generators/`. This greatly reduces cognitive load when modifying suggestion logic and makes the codebase easier for AI agents to process contextually. Shared types and constants were moved to `suggestionEngineTypes.ts` and `constants.ts` to prevent circular dependencies.
 - **Biome Strictness**: Biome linting correctly flagged the need to re-sort imports after moving everything out. Using `pnpm check:fix` effectively applies auto-fixes without manually resolving line-by-line diffs.
 - **Vitest Browser Dependency**: Tests may unexpectedly fail with `browserType.launch: Executable doesn't exist` when `pnpm test` triggers browser tests on a clean image without playwright binaries downloaded. The resolution is to always run `pnpm exec playwright install` if this test error occurs.
+- **Monolithic functions are hard to parse:** `suggestionEngine.ts` contained a massive `generateSuggestions` function spanning hundreds of lines. Extracting the HM/Item dependency checking logic (`hasHeadbutt`, `hasSurf`, etc.) and the post-processing filter loop into a dedicated utility (`encounterTools.ts`) significantly reduced the cognitive load required to understand the core engine logic.
+- **Mutation-in-place:** The `generateCatchSuggestions` and `suggestionEngine` heavily rely on mutating arrays (`suggestions.push()`) and Sets (`localPids.add()`) in place. While this is an optimization to avoid intermediate allocations, it makes data flow harder to trace without careful reading. The newly extracted `filterSuggestionsByMissingTools` follows this pattern.
+- **Separation of concerns:** Moving the `PlayerTools` interface and logic out allows other modules to potentially reuse the tool-checking logic without importing the entire suggestion engine.
